#11737. NTIS update set ishVersion only once.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@17043 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ admin/src/main/java/ish/oncourse/admin/services/ntis/NTISResult.java
@@ -1,31 +1,32 @@
 package ish.oncourse.admin.services.ntis;
 
 public class NTISResult {
+
 	private Class<?> type;
-	private int numberOfUpdated;
-	private int numberOfNew;
-	
+	private long numberOfUpdated;
+	private long numberOfNew;
+
 	public Class<?> getType() {
 		return type;
 	}
-	
+
 	public void setType(Class<?> type) {
 		this.type = type;
 	}
-	
-	public int getNumberOfNew() {
+
+	public long getNumberOfNew() {
 		return numberOfNew;
 	}
-	
-	public void setNumberOfNew(int numberOfNew) {
+
+	public void setNumberOfNew(long numberOfNew) {
 		this.numberOfNew = numberOfNew;
 	}
-	
-	public int getNumberOfUpdated() {
+
+	public long getNumberOfUpdated() {
 		return numberOfUpdated;
 	}
-	
-	public void setNumberOfUpdated(int numberOfUpdated) {
+
+	public void setNumberOfUpdated(long numberOfUpdated) {
 		this.numberOfUpdated = numberOfUpdated;
 	}
 }
